+++ packages/youtube-mute-skip-ads/src/utils.ts
@@ -1,11 +1,13 @@
-import { error, warn } from "./log";
+import { debug, debugging, error, warn } from "./log";
 
 export const playerId = "movie_player";
 export const videoSelector = "#movie_player video";
 export const muteButtonSelector =
   ":is(.ytp-mute-button, ytdDesktopShortsVolumeControlsMuteIconButton, ytmusic-player-bar tp-yt-paper-icon-button.volume)";
+export const shortsRendererSelector = "ytd-reel-video-renderer";
 export const shortsPlayerId = "shorts-player";
 export const shortsVideoSelector = "#shorts-player video";
+export const shortsUpButtonSelector = "#navigation-button-up button";
 export const shortsDownButtonSelector = "#navigation-button-down button";
 
 export type PlayerState = {
@@ -42,6 +44,9 @@ export function getMuteButton(): HTMLElement | null {
   return getHTMLElementBySelector(muteButtonSelector);
 }
 
+export function getShortsUpButton(): HTMLElement | null {
+  return getHTMLElementBySelector(shortsUpButtonSelector);
+}
 export function getShortsDownButton(): HTMLElement | null {
   return getHTMLElementBySelector(shortsDownButtonSelector);
 }
@@ -61,6 +66,33 @@ export function getHTMLElementBySelector(selector: string): HTMLElement | null {
   }
   error("Failed to find", JSON.stringify(selector));
   return null;
+}
+
+/**
+ * Given any element within an `ytd-reel-video-renderer`, such as the
+ * `#shorts-player` descendant, or its `video` descendant, or the
+ * `ytd-reel-video-renderer` element itself, return the parent of the
+ * `ytd-reel-video-renderer` element.
+ *
+ * This is useful for determining the order of adjacent shorts.
+ *
+ * Returns `null` if no such ancestor is found.
+ */
+export function getShortsParentElement(
+  elemWithinShort: Element,
+): Element | null {
+  const shortsRenderer = elemWithinShort.closest(shortsRendererSelector);
+  if (shortsRenderer == null) {
+    if (debugging) {
+      debug(
+        `Failed to find the ${shortsRendererSelector} ancestor of:`,
+        elemWithinShort.cloneNode(true),
+      );
+    }
+    return null;
+  }
+
+  return shortsRenderer.parentElement;
 }
 
 export function callMoviePlayerMethod(